pyproject.toml
[project]
name = "auto-gpt-work-main"
version = "0.1.0"
description = "从零到一实现一个Agent"
authors = [
    {name = "hedefu",email = "hedefu999@163.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "colorama (==0.4.6)",
    "langchain (==0.3.27)",
    "langchain-core (==0.3.72)",
    "langchain-community (==0.3.27)",
    "pandas (==2.3.2)",
    "pydantic (==2.11.7)",
    "python-dotenv (==1.1.1)",
    "chroma-hnswlib (==0.7.2)",
    "chromadb (==1.1.0)",
    "tiktoken (==0.7.0)",
    "openpyxl (==3.1.5)",
    "pymupdf (==1.24.7)",
    "onnxruntime (==1.19.2)",
    "zai-sdk (==0.0.3.3)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

